+++ apis/database/__init__.py
@@ -1,1 +1,1 @@
-from . import base, user
+from . import base, user, contents
